+++ CHANGELOG.md
@@ -5,6 +5,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.28.0] - 2020-07-02
+
+### Addded
+
+- Copy and link files from the ExtData folder when importing relANNIS.
+- Map `resolver_vis_map.annis`, `example_queries.annis` and `corpus.properties` from relANNIS files 
+  to a new unified corpus configuration stored as [TOML]() file. This corpus configuration
+  is also exported to GraphML.
+- Export and import ZIP files containing multiple corpora.
+
+### Removed
+
+- Removed Brotli support: use the ZIP file export instead
+
 ## [0.27.0] - 2020-06-08
 
 ### Changed